Compiler buildscript: verify dlnd'd files and reuse existing files
The buildscript now recognizes existing gcc and binutils folders and
will not download those them if not needed. If a download is needed,
the downloaded files will be verified with a static SHA-256 hash stored
in the repo to prevent corruption and as a basic security measure. If
real security is desired, the real digital signatures from GNU are available.